requirements.txt
# Python 3.13.12

# aws_sync.py
pandas>=2.2.0
pyarrow>=16.0.0
boto3>=1.34.0

# analytics_dashboard.py + script.py
psycopg2-binary>=2.9.9
seaborn>=0.13.0
matplotlib>=3.9.0
